requirements.txt
asgiref==3.2.7
automium==0.2.6
automium-web==0.1.1
certifi==2020.4.5
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-allauth==0.41.0
django-extensions==2.2.9
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
Flask==1.1.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-SSE==0.2.1
Flask-WTF==0.14.3
gunicorn==20.0.4
http-ece==1.1.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
legacy==0.1.6
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.2
oauthlib==3.1.0
ok==0.0.0
passlib==1.7.2
psycopg2==2.8.4
py-vapid==1.7.0
pycparser==2.20
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywebpush==1.10.1
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
Werkzeug==1.0.0
whitenoise==5.0.1
WTForms==2.2.1
